Move task discovery into tasks.ts
discoverTasks and rescanTasks follow the preference state that moved last
commit: every input they read — taskPrefs, isTrusted, applyRunner,
lastRunnableById, hiddenIds — is now in tasks.ts, so they were reaching back
across the boundary for all of it.

Done for the same reason as that commit: the Run picker and the project tasks
panel are next out, and both call discovery. Leaving it here would have made it
two more hooks on each.

Byte-identical apart from `export`. tasks.ts now imports invoke and dlog, which
is the module's first backend call — the rest of it is still pure.
