+++ src/main.ts
@@ -58,7 +58,7 @@ import {
   dormantBusy, dormantRows, extWorking, GCLASS, GLYPH, groupBody, taskStateText,
 } from "./sidebarview";
 import {
-  applyInputs, applyRunner, execCmd, exitWaiters, hiddenIds, isTrusted,
+  applyInputs, discoverTasks, execCmd, exitWaiters, hiddenIds, rescanTasks,
   lastRunnableById, launchWithDeps, pinnedIds, PROVIDER_LABEL, rememberedInput,
   rememberInput, RUNNERS, runnerFor, setTaskLauncher, setTaskLogger, setTaskRepaint,
   setTaskToast, stopRuleBlocked, stopRules, taskPrefs, taskRunner, toggleHidden,
@@ -1225,34 +1225,8 @@ async function rerunTask(s: Sess) {
   await launchTask(spec, project, { colorKey, worktree, branch });
 }
 
-// `trusted` is what lets the backend shell out to `just --dump` — which evaluates
-// the justfile. It takes all three of: the global toggle, the provider being on,
-// and this specific folder being one the user chose.
-async function discoverTasks(workdir: string, colorKey = workdir, includeHidden = false): Promise<Runnable[]> {
-  const trusted = taskPrefs.introspect && taskPrefs.providers.includes("just") && (isTrusted(workdir) || isTrusted(colorKey));
-  try {
-    const raw = (await invoke<Runnable[]>("discover_runnables", { workdir, trusted }))
-      .filter((r) => taskPrefs.providers.includes(r.source as Provider));
-    // Swap in the project's runner override before anything caches the result, so a
-    // re-run months later uses the same runner the picker showed.
-    const all = applyRunner(raw, colorKey);
-    // Resolve dependsOn against everything discovered, hidden or not — hiding a
-    // task from the picker shouldn't quietly break another task that needs it.
-    for (const r of all) lastRunnableById.set(r.id, r);
-    const hid = hiddenIds(colorKey);
-    return includeHidden ? all : all.filter((r) => !hid.includes(r.id));
-  } catch (e) {
-    dlog("warn", `discover failed (${workdir}): ${e}`);
-    return [];
-  }
-}
-
-// Drop the backend's cached parse for this project so the next discover re-reads
-// from disk. The stamp already catches edits to files Episko reads; this is the
-// escape hatch for what it can't see — a file an introspector imports itself.
-async function rescanTasks(workdir: string) {
-  await invoke("rescan_runnables", { workdir }).catch((e) => dlog("warn", `rescan: ${e}`));
-}
+// Discovery (discoverTasks / rescanTasks) moved to ./tasks — every input they read
+// now lives there.
 
 // ---------- the inputs prompt ----------
 // A task declaring ${input:…} collects its values before anything runs. Discovery
